+++ app/src/main/java/io/vn/nguyenduck/blocktopograph/activity/StartActivity.java
@@ -4,15 +4,14 @@ import static android.Manifest.permission.READ_EXTERNAL_STORAGE;
 import static android.Manifest.permission.WRITE_EXTERNAL_STORAGE;
 import static android.content.pm.PackageManager.PERMISSION_GRANTED;
 import static android.provider.Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION;
-import static io.vn.nguyenduck.blocktopograph.Constants.BOGGER;
-import static io.vn.nguyenduck.blocktopograph.Constants.MINECRAFT_APP_ID;
 import static io.vn.nguyenduck.blocktopograph.Constants.SHIZUKU_PACKAGE_NAME;
-import static io.vn.nguyenduck.blocktopograph.Constants.WORLDS_FOLDER;
-import static io.vn.nguyenduck.blocktopograph.utils.Utils.buildAndroidDataDir;
-import static io.vn.nguyenduck.blocktopograph.utils.Utils.buildMinecraftDataDir;
+import static io.vn.nguyenduck.blocktopograph.shell.Runner.isRooted;
 import static io.vn.nguyenduck.blocktopograph.utils.Utils.isAndroid11Up;
 
+import android.app.AlertDialog;
 import android.content.Intent;
+import android.content.pm.ActivityInfo;
+import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
 
@@ -21,12 +20,8 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.Arrays;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.logging.Level;
 
 import io.vn.nguyenduck.blocktopograph.R;
-import io.vn.nguyenduck.blocktopograph.file.BFile;
 import rikka.shizuku.Shizuku;
 
 public class StartActivity extends AppCompatActivity {
@@ -34,47 +29,22 @@ public class StartActivity extends AppCompatActivity {
     private final int STORAGE_PERMISSION_CODE = 0x7832;
     private final int SHIZUKU_PERMISSION_CODE = 0xbfde;
 
-    private boolean StoragePermission;
-    private boolean ShizukuPermission;
+    private boolean StoragePermission = false;
+    private boolean ShizukuPermission = false;
     private boolean ShizukuInstalled = false;
-
-    private final ExecutorService executor = Executors.newWorkStealingPool();
-
-    private final Shizuku.OnBinderReceivedListener BINDER_RECEIVED_LISTENER = () -> {
-        if (Shizuku.isPreV11()) showNotSupportedShizukuVersion();
-    };
-
-    private final Shizuku.OnRequestPermissionResultListener REQUEST_PERMISSION_RESULT_LISTENER = this::onRequestPermissionResult;
-
-    private void showNotSupportedShizukuVersion() {
-        BOGGER.info("Not support shizuku version lower than 11");
-    }
-
-    private void onRequestPermissionResult(int requestCode, int result) {
-        if (requestCode == SHIZUKU_PERMISSION_CODE)
-            ShizukuPermission = result == PERMISSION_GRANTED;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == STORAGE_PERMISSION_CODE) {
-            StoragePermission = grantResults.length > 0 &&
-                    Arrays.stream(grantResults).allMatch(v -> v == PERMISSION_GRANTED);
-        }
-    }
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.start_activity);
 
-        Shizuku.addBinderReceivedListenerSticky(BINDER_RECEIVED_LISTENER);
         Shizuku.addRequestPermissionResultListener(REQUEST_PERMISSION_RESULT_LISTENER);
 
-        StoragePermission = hasFileAccessPermission();
-        ShizukuInstalled = hasInstalledShizuku();
-        ShizukuPermission = hasShizukuPermission();
+        // setup for auto rotated screen on sensor
+        setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_SENSOR);
+    }
+
+    private void startMainActivity() {
         startActivity(new Intent(this, MainActivity.class));
         finish();
     }
@@ -82,21 +52,27 @@ public class StartActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        if (StoragePermission && ShizukuPermission) {
-            try {
-                var p = buildAndroidDataDir(MINECRAFT_APP_ID);
-                var f = new BFile(buildMinecraftDataDir(p, WORLDS_FOLDER));
-                BOGGER.info(Arrays.toString(f.listDirs()));
-            } catch (Exception e) {
-                BOGGER.log(Level.SEVERE, e, () -> "");
-            }
-        } else {
-            if (!StoragePermission) requestStoragePermission();
-            if (isAndroid11Up() && !ShizukuInstalled) {
-                BOGGER.info("Shizuku Not Installed!");
-            } else if (Shizuku.pingBinder() && !ShizukuPermission) {
+
+        StoragePermission = StoragePermission || hasFileAccessPermission();
+        ShizukuInstalled = ShizukuInstalled || hasInstalledShizuku();
+        ShizukuPermission = ShizukuPermission || hasShizukuPermission();
+
+        if (!StoragePermission) requestStoragePermission();
+        if (isAndroid11Up()) {
+            if (isRooted()) {
+                startMainActivity();
+            } else if (!ShizukuInstalled) {
+                showShizukuAppNeeded();
+            } else if (!Shizuku.pingBinder()) {
+                showNotRunningShizuku();
+            } else if (!ShizukuPermission) {
                 Shizuku.requestPermission(SHIZUKU_PERMISSION_CODE);
             }
+            if (StoragePermission && ShizukuPermission) {
+                startMainActivity();
+            }
+        } else if (StoragePermission) {
+            startMainActivity();
         }
     }
 
@@ -104,7 +80,6 @@ public class StartActivity extends AppCompatActivity {
     protected void onDestroy() {
         super.onDestroy();
 
-        Shizuku.removeBinderReceivedListener(BINDER_RECEIVED_LISTENER);
         Shizuku.removeRequestPermissionResultListener(REQUEST_PERMISSION_RESULT_LISTENER);
     }
 
@@ -118,8 +93,7 @@ public class StartActivity extends AppCompatActivity {
     }
 
     private boolean hasShizukuPermission() {
-        if (!Shizuku.pingBinder()) return false;
-        return Shizuku.checkSelfPermission() == PERMISSION_GRANTED;
+        return Shizuku.pingBinder() && Shizuku.checkSelfPermission() == PERMISSION_GRANTED;
     }
 
     private boolean hasFileAccessPermission() {
@@ -143,6 +117,54 @@ public class StartActivity extends AppCompatActivity {
                     new String[]{READ_EXTERNAL_STORAGE, WRITE_EXTERNAL_STORAGE},
                     STORAGE_PERMISSION_CODE
             );
+        }
+    }
+
+    private void showShizukuAppNeeded() {
+        new AlertDialog.Builder(this)
+                .setTitle("Shizuku Not Installed!")
+                .setMessage("You need to install shizuku first")
+                .setPositiveButton("Open Google Play Store", (d, w) -> showShizukuInMarket())
+                .setNegativeButton("Exit", (d, w) -> System.exit(0))
+                .setCancelable(false)
+                .show();
+    }
+
+    private void showNotRunningShizuku() {
+        new AlertDialog.Builder(this)
+                .setTitle("Shizuku Not Running!")
+                .setMessage("You need to start shizuku and run it first")
+                .setPositiveButton("Open Shizuku", (d, w) -> openShizuku())
+                .setNegativeButton("Exit", (d, w) -> System.exit(0))
+                .setCancelable(false)
+                .show();
+    }
+
+    private void showShizukuInMarket() {
+        try {
+            startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + SHIZUKU_PACKAGE_NAME)));
+        } catch (Exception ignored) {
+            startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=" + SHIZUKU_PACKAGE_NAME)));
+        }
+    }
+
+    private void openShizuku() {
+        startActivity(getPackageManager().getLaunchIntentForPackage(SHIZUKU_PACKAGE_NAME));
+    }
+
+    private final Shizuku.OnRequestPermissionResultListener REQUEST_PERMISSION_RESULT_LISTENER = this::onRequestPermissionResult;
+
+    private void onRequestPermissionResult(int requestCode, int result) {
+        if (requestCode == SHIZUKU_PERMISSION_CODE)
+            ShizukuPermission = result == PERMISSION_GRANTED;
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == STORAGE_PERMISSION_CODE) {
+            StoragePermission = grantResults.length > 0 &&
+                    Arrays.stream(grantResults).allMatch(v -> v == PERMISSION_GRANTED);
         }
     }
 }