+++ release_notes.md
@@ -4,3 +4,10 @@
 - Removed old main activity design
 - Added new design for better performance
 - Added game engine library
+
+### 18Oct24
+- Added a way to check rooted
+- Added show shizuku not running required dialog when start app
+
+### 19Oct24
+- Auto copy games from Android/data to /sdcard/games folder and back